Fix remapped shader not found
Fixes #6 (Warn: R_RemapShader: new shader textures/thunder/blend not
found). The visible consequence of this bug was that one of the lighting
bolts is displayed at the beginning of the map.

Beyond that, the mistake didn't have further consequences because it was
canceled out by an engine bug. When the replacement shader is not found,
R_RemapShader uses a "defaulted" shader, which are never supposed to be
used for rendering, as the replacement. The defaulted shader has 0
stages so it renders nothing, which happens to be what
textures/thunder/blend is supposed to do.